Break apart complex UI views
